+++ Testing2/TestingStaff.csproj
@@ -51,7 +51,6 @@
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="tststaff.cs" />
-    <Compile Include="clsStaff.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
